backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
Pillow>=10.2.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
